feat: switch to `lol-html`
cloudflare/html-rewriter-wasm is outdated with no significant updates
since 2022. Meanwhile, the underlying `lol-html` library has been
actively maintained and has even reached 2.0.

The JS binding for `lol-html` was published to NPM from an unofficial
fork, but it has caught up with upstream so we can use it now.

Initial testing shows that it can be a drop-in replacement for
`cloudflare/html-rewriter-wasm`.

Resolves #1